+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Bugfixes
 
 ### Added
+* Add validation for icons and screenshots. [#537](https://github.com/elastic/package-registry/pull/537)
 
 ### Deprecated